requirements.txt
streamlit>=1.34.0
pandas
plotly
google-generativeai
python-dotenv
SQLAlchemy
psycopg2-binary
